+++ instrumentation/jdbc/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jdbc/DataSourceInstrumentation.java
@@ -1,24 +1,15 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
 package io.opentelemetry.javaagent.instrumentation.jdbc;
 
 import static io.opentelemetry.api.trace.SpanKind.CLIENT;
 import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.implementsInterface;
 import static io.opentelemetry.javaagent.instrumentation.jdbc.DataSourceTracer.tracer;
-import static java.util.Collections.singletonList;
 import static java.util.Collections.singletonMap;
 import static net.bytebuddy.matcher.ElementMatchers.named;
 
-import com.google.auto.service.AutoService;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.context.Scope;
-import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import io.opentelemetry.javaagent.instrumentation.api.Java8BytecodeBridge;
-import java.util.List;
 import java.util.Map;
 import javax.sql.DataSource;
 import net.bytebuddy.asm.Advice;
@@ -26,32 +17,16 @@ import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
-@AutoService(InstrumentationModule.class)
-public class JdbcDataSourceInstrumentationModule extends InstrumentationModule {
-  public JdbcDataSourceInstrumentationModule() {
-    super("jdbc-datasource");
+public class DataSourceInstrumentation implements TypeInstrumentation {
+  @Override
+  public ElementMatcher<TypeDescription> typeMatcher() {
+    return implementsInterface(named("javax.sql.DataSource"));
   }
 
   @Override
-  public List<TypeInstrumentation> typeInstrumentations() {
-    return singletonList(new DataSourceInstrumentation());
-  }
-
-  @Override
-  public boolean defaultEnabled() {
-    return false;
-  }
-
-  public static class DataSourceInstrumentation implements TypeInstrumentation {
-    @Override
-    public ElementMatcher<TypeDescription> typeMatcher() {
-      return implementsInterface(named("javax.sql.DataSource"));
-    }
-
-    @Override
-    public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-      return singletonMap(named("getConnection"), GetConnectionAdvice.class.getName());
-    }
+  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
+    return singletonMap(named("getConnection"),
+        DataSourceInstrumentation.class.getName() + "$GetConnectionAdvice");
   }
 
   public static class GetConnectionAdvice {
